+++ package.json
@@ -9,19 +9,19 @@
     "start": "next start"
   },
   "dependencies": {
-    "@ant-design/cssinjs": "^1.16.1",
-    "@ant-design/pro-components": "^2.6.28",
-    "@prisma/client": "^5.1.1",
+    "@ant-design/cssinjs": "^1.17.2",
+    "@ant-design/pro-components": "^2.6.32",
+    "@prisma/client": "^5.5.1",
     "@types/node": "20.4.2",
     "@types/react": "18.2.15",
     "@types/react-dom": "18.2.7",
     "@vercel/edge-config": "^0.2.1",
-    "@vercel/postgres": "^0.4.1",
-    "@vercel/postgres-kysely": "^0.4.1",
-    "ai": "^2.2.11",
-    "antd": "^5.8.2",
+    "@vercel/postgres": "^0.4.2",
+    "@vercel/postgres-kysely": "^0.4.2",
+    "ai": "^2.2.18",
+    "antd": "^5.10.2",
     "autoprefixer": "10.4.14",
-    "axios": "^1.5.0",
+    "axios": "^1.5.1",
     "crypto-browserify": "^3.12.0",
     "dotenv": "^16.3.1",
     "eslint": "8.45.0",
@@ -32,10 +32,10 @@
     "next": "13.4.10",
     "next-auth": "^4.23.1",
     "node-fetch": "^3.3.2",
-    "openai": "^4.0.0",
+    "openai": "^4.13.0",
     "openai-proxy": "^0.0.1",
     "perf_hooks": "^0.0.1",
-    "pg": "^8.11.2",
+    "pg": "^8.11.3",
     "postcss": "8.4.26",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -44,14 +44,14 @@
     "sse": "^0.0.8",
     "stream": "^0.0.2",
     "stream-to-blob": "^2.0.1",
-    "styled-components": "^6.0.7",
+    "styled-components": "^6.1.0",
     "tailwindcss": "3.3.3",
     "typescript": "5.2.2",
-    "uuid": "^9.0.0",
-    "ws": "^8.14.1"
+    "uuid": "^9.0.1",
+    "ws": "^8.14.2"
   },
   "devDependencies": {
-    "@types/uuid": "^9.0.3",
-    "prettier": "^3.0.1"
+    "@types/uuid": "^9.0.6",
+    "prettier": "^3.0.3"
   }
 }
